Remove dependency on os-lib (#16)
It is a binary compatibility concern (via transitive dependency on
geny). This would be okay since we can create a fat jar and shade
dependencies, but trying to shade package os.** seems to cause issues
with loading system property "os.name" which we need to use the correct
binary for the user's system.

We aren't gaining much by using it and it causes problems, so remove the
dependency.