Change to "mostly-statically linked" jar (#17)
Most dependencies (including transitive ones) are now assembled into the
published jar. These assembled dependencies are also shaded (ie. their
packages are changed) so that they do not collide with the same
dependency pulled in dynamically. This is effectively the equivalent of
static linking. Only Scala library dependencies are kept dynamic since
this project is intended to be used by Scala projects and those will be
needed anyway. This reduces the size of the jar although it is still
fairly big.